requirements.txt
altgraph==0.17
cached-property==1.5.2
certifi==2020.12.5
click==7.1.2
docutils==0.17.1
eric-ide==20.2
future==0.18.2
h5py==3.2.1
importlib-metadata==4.0.1
iso8601==0.1.14
Markdown==3.3.4
numpy==1.20.3
opencv-python==4.5.2.52
pefile==2019.4.18
pyinstaller==4.3
pyinstaller-hooks-contrib==2021.1
PyQt5==5.15.4
pyqt5-plugins==5.15.4.2.1.0
PyQt5-Qt5==5.15.2
PyQt5-sip==12.8.1
pyqt5-tools==5.15.4.3.0.3
PyQtWebEngine==5.15.4
PyQtWebEngine-Qt5==5.15.2
pyserial==3.5
python-dotenv==0.17.1
pyueye==4.90.0.0
pywin32==300
pywin32-ctypes==0.2.0
PyYAML==5.4.1
QScintilla==2.12.1
qt5-applications==5.15.2.2.1
qt5-tools==5.15.2.1.0.1
scipy==1.6.3
tqdm==4.60.0
typing-extensions==3.10.0.0
wincertstore==0.2
zipp==3.4.1
